chore(refactoring): removed all 'views' from extensions
