--- extensions/lite/views/RulesChecker/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const RulesChecker = props => props.children
-
-export default RulesChecker
